requirements.txt
aioice==0.10.1
aiortc==1.14.0
annotated-types==0.7.0
anyio==4.11.0
av==16.0.1
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
dnspython==2.8.0
fastapi==0.117.1
google-crc32c==1.7.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
packaging==25.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
